+++ www/static/js/dw/chart/visualize/options.js
@@ -1,8 +1,8 @@
 
 define(
-    ['./options/initCustomColors', './options/syncVisOptions'],
+    ['./options/initCustomColors', './options/syncVisOptions', './options/unsyncVisOptions'],
 
-function(initCustomColors, syncVisOptions) {
+function(initCustomColors, syncVisOptions, unsyncVisOptions) {
 
     var _chart, _vis;
 
@@ -16,6 +16,17 @@ function(initCustomColors, syncVisOptions) {
             syncVisOptions(_vis, _chart);
             $('.select-row').hide();
             initCustomColors(_chart);
+
+            $('.vis-option-help').tooltip({
+                placement: 'left',
+                html: true,
+                trigger: 'hover focus click',
+                container: 'body'
+            });
+        },
+
+        reset: function() {
+            unsyncVisOptions(_vis, _chart);
         }